Replace fixtures with factories for LobAddress and Check (#3007)
This PR started out as replacing LobAddress fixtures with factories for good, since it has burned us in a few PRs now. It was trickier than other models because the LobAddress model makes remote API calls to Lob's service. My initial approach was to provide Faker generated address data, which actually made the factory valid because it would call Lob's development service with the fake address. However, it turned out to be pretty easy to turn off any remote API calls to Lob in the test environment, and since none of the tests affected here use lob_id, we were good.

After that I replaced use of LobAddress fixtures everywhere with the factory and that also led me to Checks, which were a lot easier to replace. So this PR contains fixes to both.

* Add valid LobAddress factory
* Delete lob_addresses fixture from lob_address_spec
* Skip calling the real Lob service in test

The only real thing we seem to do with calling Lob is storing the
lob_id. We can consider refactoring this to be a more fully fledged
mock/fake if there's a test for that.

* Update CheckService::Create spec with lob address factory

This test was introduced in https://github.com/hackclub/bank/pull/2804
and fixtures were used as a stopgap.

* Clean up check_spec with factory

* Remove checks fixture

+++ app/models/lob_address.rb
@@ -59,6 +59,8 @@ class LobAddress < ApplicationRecord
   end
 
   def create_lob_address
+    return if Rails.env.test?
+
     lob_address = LobService.instance.add_address(
       description,
       name,
@@ -74,6 +76,8 @@ class LobAddress < ApplicationRecord
   end
 
   def update_lob_address
+    return if Rails.env.test?
+
     lob_address = LobService.instance.update_address(
       lob_id,
       description,
@@ -90,6 +94,8 @@ class LobAddress < ApplicationRecord
   end
 
   def destroy_lob_address
+    return if Rails.env.test?
+
     LobService.instance.delete_address(lob_id)
   end
